+++ cpp/results.hpp
@@ -1,0 +1,19 @@
+#pragma once
+// from http://kevingong.com/Polyominoes/Enumeration.html
+uint64_t results[] = {
+    1,
+    1,
+    2,
+    8,
+    29,
+    166,
+    1023,
+    6922,
+    48311,
+    346543,
+    2522522,
+    18598427,
+    138462649,
+    1039496297,
+    7859514470,
+    59795121480};
